Bug_5081 - Refactor few functions from eval.c
